fix: 컨트롤러 예외 처리 수정
