+++ src/main/java/roomescape/reservation/presentation/MemberReservationController.java
@@ -1,14 +1,18 @@
 package roomescape.reservation.presentation;
 
+import jakarta.servlet.http.HttpServletRequest;
 import java.net.URI;
+import java.time.format.DateTimeParseException;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 import roomescape.auth.login.presentation.dto.LoginMemberInfo;
 import roomescape.auth.login.presentation.dto.annotation.LoginMember;
+import roomescape.common.exception.handler.dto.ExceptionResponse;
 import roomescape.member.presentation.dto.MyReservationResponse;
 import roomescape.reservation.presentation.dto.ReservationRequest;
 import roomescape.reservation.presentation.dto.ReservationResponse;
@@ -23,7 +27,7 @@ public class MemberReservationController {
         this.reservationService = reservationService;
     }
 
-    @PostMapping
+    @PostMapping("/reservations")
     public ResponseEntity<ReservationResponse> createReservation(
         @RequestBody final ReservationRequest request,
         @LoginMember final LoginMemberInfo memberInfo)
@@ -37,5 +41,13 @@ public class MemberReservationController {
         List<MyReservationResponse> response = reservationService.getMemberReservations(loginMemberInfo);
 
         return ResponseEntity.ok().body(response);
+    }
+
+    @ExceptionHandler(value = DateTimeParseException.class)
+    public ResponseEntity<ExceptionResponse> noMatchDateType(final HttpServletRequest request) {
+        ExceptionResponse exceptionResponse = new ExceptionResponse(
+            400, "[ERROR] 요청 날짜 형식이 맞지 않습니다.", request.getRequestURI()
+        );
+        return ResponseEntity.badRequest().body(exceptionResponse);
     }
 }
